+++ changelog.md
@@ -2,6 +2,8 @@
 
 **Version 0.2.0**
 
+* Fixed issues with reading mixed `jpg` and `png` images. 
+
 * Added `reshape_images` and `resize_images`. These are helper functions that may be useful when plotting
 or restoring original image size.